refactor(connections): extract sections into components
